revise to new ztunnel command
